Initial refactor from Manage Poitions to Manage Redactors
